+++ angular/projects/wcs-angular/src/lib/directives/select-value-accessor.directive.ts
@@ -1,6 +1,5 @@
-import { Directive, ElementRef, Injector } from '@angular/core';
+import { Directive, ElementRef, HostListener, Injector } from '@angular/core';
 import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
-import { SelectChangeEventDetail } from '../../../../../../dist/types/components/select/select-interface';
 
 @Directive({
   /* tslint:disable-next-line:directive-selector */
@@ -15,33 +14,27 @@ import { SelectChangeEventDetail } from '../../../../../../dist/types/components
 })
 // tslint:disable-next-line:directive-class-suffix
 export class SelectValueAccessor implements ControlValueAccessor {
-  private value;
   private onChange: (value: any) => void = () => {};
   private onTouched: () => void = () => {};
 
   constructor(protected injector: Injector, protected el: ElementRef) {
-    this.el.nativeElement.addEventListener('wcsChange', (event: CustomEvent<SelectChangeEventDetail>) => {
-      this.onChange(event.detail.value);
-    });
+  }
 
-
+  @HostListener('wcsChange', ['$event.detail.value'])
+  _handleInputEvent(value: any): void {
+      this.onChange(value);
   }
 
   // tslint:disable-next-line:typedef
   writeValue(value) {
-    this.value = value;
-    this.el.nativeElement.value = this.value;
+    this.el.nativeElement.value = value;
   }
 
-  // tslint:disable-next-line:typedef
-  registerOnChange(value) {
-    this.onChange = value;
+  registerOnChange(fn: (_: any) => void): void {
+    this.onChange = fn;
   }
 
-  /**
-   * Not implemented for now
-   */
-  // tslint:disable-next-line:typedef
-  registerOnTouched(fn) {
+  registerOnTouched(fn: any): void {
+    this.onTouched = fn;
   }
 }